utils.sh: listen on IPv4 and IPv6 by default
Using 0.0.0.0 as default listen IP causes the container ports to only be
reachable on IPv4. Change the LISTEN_IP variable to append : if the
variable is set, so that when the user wants to override LISTEN_IP, it
will be in the correct format for the publish argument. If it is not
set, we don't specify an address in the publish argument at all, and
Docker proxy will bind a socket using IPv4 and IPv6 wildcard addresses.